requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-cors-headers==1.1.0
django-filters==0.1.0
django-rest-swagger==0.3.2
django-toolbelt==0.0.1
djangorestframework==3.1.3
docopt==0.4.0
gunicorn==19.3.0
jws==0.1.3
mandrill==1.0.57
Markdown==2.6.2
psycopg2==2.6
python-jwt==1.0.3
PyYAML==3.11
requests==2.7.0
static3==0.6.1
